Setup codebase for GC implementation
this is a first PR to refactor the codebase a bit so that it is amenable
to the addition of a GC.

The main idea is that that Alloc will at some point keep track of all
the Values it has allocated (so it cannot allocate any T anymore, it has
to be a T that can be wrapped in a Value type). This is necessary so
that we can track the values that have been allocated.

I think this is necessary even with a copying GC, which seems to be
possibly the preferred solution as per
https://github.com/maximecb/plush/issues/25, as it may be necessary to
allow 0 copy message sending across actors.

The message allocators, in my mind, will still be the bump where the
msgs live, but once a message is received it will be marked as received.
At that point, it would act as if it's a forever root. The received
message will also be tracked (shallowly, meaning just the metadata) by
the actor GC and once it's no longer in use, it will somehow alert the
message allocator to remove it from the roots. This is why I think the
allocators need to track all values they allocate.

The Alloc also takes roots, which is the live values that can be derived
by the actor state. I've added some helpers functions to actor so that
it's easier to allocate if you have an actor.

I've also made it so that we know if we're in a message allocator or not
inside the allocator, so that in future, we can create the right
metadata depending if we're in a message allocator or in an actor
allocator.

One annoying part of this whole approach is that we now need an actor at
codegen, not just an allocator, as I've essentially made the actor the
thing that is responsible for providing the roots list to the allocator.

All in all, I think that this is a decent start, which should allow
follow up changes to add a GC, of any kind (hopefully), to the VM. I
could also be very wrong!