Simplify due to copying GC design
after maxime's very clear review comment, it's clear that we won't need
to keep track of values in the allocator. Thus, i've simplified a bit
the code, the allocator is now fallible (when it's full) and it's the
actors responsibility to clear it up.

I've also introduced a new MsgAlloc struct which will, in the future,
block on the gc_cond, when allocations on the message allocator fail.
The idea is that all threads trying to push to the message allocator
will just block till the actor has run has gc which has also cleaned up
the message allocator, at that point it will notify_all on the condvar.